+++ adcp-20220101/alibabacloud_adcp20220101/client.py
@@ -21,7 +21,6 @@ class Client(OpenApiClient):
         config: open_api_models.Config,
     ):
         super().__init__(config)
-        self._signature_algorithm = 'v2'
         self._endpoint_rule = 'central'
         self._endpoint_map = {
             'cn-beijing': 'adcp.cn-beijing.aliyuncs.com',
@@ -172,6 +171,8 @@ class Client(OpenApiClient):
             body['Profile'] = request.profile
         if not UtilClient.is_unset(request.region_id):
             body['RegionId'] = request.region_id
+        if not UtilClient.is_unset(request.resource_group_id):
+            body['ResourceGroupID'] = request.resource_group_id
         if not UtilClient.is_unset(request.v_switches):
             body['VSwitches'] = request.v_switches
         if not UtilClient.is_unset(request.vpc_id):
@@ -220,6 +221,8 @@ class Client(OpenApiClient):
             body['Profile'] = request.profile
         if not UtilClient.is_unset(request.region_id):
             body['RegionId'] = request.region_id
+        if not UtilClient.is_unset(request.resource_group_id):
+            body['ResourceGroupID'] = request.resource_group_id
         if not UtilClient.is_unset(request.v_switches):
             body['VSwitches'] = request.v_switches
         if not UtilClient.is_unset(request.vpc_id):
@@ -826,6 +829,8 @@ class Client(OpenApiClient):
         query = {}
         if not UtilClient.is_unset(request.profile):
             query['Profile'] = request.profile
+        if not UtilClient.is_unset(request.resource_group_id):
+            query['ResourceGroupId'] = request.resource_group_id
         req = open_api_models.OpenApiRequest(
             query=OpenApiUtilClient.query(query)
         )
@@ -854,6 +859,8 @@ class Client(OpenApiClient):
         query = {}
         if not UtilClient.is_unset(request.profile):
             query['Profile'] = request.profile
+        if not UtilClient.is_unset(request.resource_group_id):
+            query['ResourceGroupId'] = request.resource_group_id
         req = open_api_models.OpenApiRequest(
             query=OpenApiUtilClient.query(query)
         )
@@ -1604,6 +1611,14 @@ class Client(OpenApiClient):
         tmp_req: adcp_20220101_models.GrantUserPermissionsRequest,
         runtime: util_models.RuntimeOptions,
     ) -> adcp_20220101_models.GrantUserPermissionsResponse:
+        """
+        @deprecated : GrantUserPermissions is deprecated, please use adcp::2022-01-01::GrantUserPermission instead.
+        
+        @param tmp_req: GrantUserPermissionsRequest
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: GrantUserPermissionsResponse
+        Deprecated
+        """
         UtilClient.validate_model(tmp_req)
         request = adcp_20220101_models.GrantUserPermissionsShrinkRequest()
         OpenApiUtilClient.convert(tmp_req, request)
@@ -1638,6 +1653,14 @@ class Client(OpenApiClient):
         tmp_req: adcp_20220101_models.GrantUserPermissionsRequest,
         runtime: util_models.RuntimeOptions,
     ) -> adcp_20220101_models.GrantUserPermissionsResponse:
+        """
+        @deprecated : GrantUserPermissions is deprecated, please use adcp::2022-01-01::GrantUserPermission instead.
+        
+        @param tmp_req: GrantUserPermissionsRequest
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: GrantUserPermissionsResponse
+        Deprecated
+        """
         UtilClient.validate_model(tmp_req)
         request = adcp_20220101_models.GrantUserPermissionsShrinkRequest()
         OpenApiUtilClient.convert(tmp_req, request)
@@ -1671,6 +1694,13 @@ class Client(OpenApiClient):
         self,
         request: adcp_20220101_models.GrantUserPermissionsRequest,
     ) -> adcp_20220101_models.GrantUserPermissionsResponse:
+        """
+        @deprecated : GrantUserPermissions is deprecated, please use adcp::2022-01-01::GrantUserPermission instead.
+        
+        @param request: GrantUserPermissionsRequest
+        @return: GrantUserPermissionsResponse
+        Deprecated
+        """
         runtime = util_models.RuntimeOptions()
         return self.grant_user_permissions_with_options(request, runtime)
 
@@ -1678,6 +1708,13 @@ class Client(OpenApiClient):
         self,
         request: adcp_20220101_models.GrantUserPermissionsRequest,
     ) -> adcp_20220101_models.GrantUserPermissionsResponse:
+        """
+        @deprecated : GrantUserPermissions is deprecated, please use adcp::2022-01-01::GrantUserPermission instead.
+        
+        @param request: GrantUserPermissionsRequest
+        @return: GrantUserPermissionsResponse
+        Deprecated
+        """
         runtime = util_models.RuntimeOptions()
         return await self.grant_user_permissions_with_options_async(request, runtime)
 
@@ -1714,8 +1751,8 @@ class Client(OpenApiClient):
             query['DeletionProtection'] = request.deletion_protection
         if not UtilClient.is_unset(request.enable_mesh):
             query['EnableMesh'] = request.enable_mesh
-        if not UtilClient.is_unset(request.mseenabled):
-            query['MSEEnabled'] = request.mseenabled
+        if not UtilClient.is_unset(request.gateway_enabled):
+            query['GatewayEnabled'] = request.gateway_enabled
         if not UtilClient.is_unset(request.monitor_enabled):
             query['MonitorEnabled'] = request.monitor_enabled
         if not UtilClient.is_unset(request.name):
@@ -1782,8 +1819,8 @@ class Client(OpenApiClient):
             query['DeletionProtection'] = request.deletion_protection
         if not UtilClient.is_unset(request.enable_mesh):
             query['EnableMesh'] = request.enable_mesh
-        if not UtilClient.is_unset(request.mseenabled):
-            query['MSEEnabled'] = request.mseenabled
+        if not UtilClient.is_unset(request.gateway_enabled):
+            query['GatewayEnabled'] = request.gateway_enabled
         if not UtilClient.is_unset(request.monitor_enabled):
             query['MonitorEnabled'] = request.monitor_enabled
         if not UtilClient.is_unset(request.name):